refactor: split the quadlet healthcheck render tests into their own file
The two x-podman-on-failure tests pushed internal/quadlet/tests/fields.rs
to 519 code lines, over the 500-line hard limit, and line-limit.yml
failed the build.

Split by topic rather than by count: healthcheck rendering — including
the Podman extensions with no Compose Specification equivalent — is a
subject of its own, and fields.rs was carrying it only because it was
where the first one landed. 481 and 41 lines.

Part of #1095
